requirements.txt
adrf==0.1.1
asgiref==3.7.2
Django==4.2.1
djangorestframework==3.14.0
Pyrogram==2.0.106
